pyproject.toml
[tool.poetry]
name = "github-user-parser"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "github_user_parser"}]

[tool.poetry.dependencies]
python = "^3.11"
SQLAlchemy = "^2.0.15"
python-dotenv = "^0.21.1"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
